+++ includes/Module/SecurityCheck/download_directory.php
@@ -12,28 +12,32 @@
   use ClicShopping\OM\CLICSHOPPING;
   use ClicShopping\OM\Registry;
 
-  class securityCheck_check_github_directory {
+  class securityCheck_download_directory
+  {
     public string $type = 'warning';
 
     public function __construct()
     {
       $CLICSHOPPING_Language = Registry::get('Language');
 
-      $CLICSHOPPING_Language->loadDefinitions('modules/security_check/check_github_directory', null, null, 'Shop');
-
-      $this->title = CLICSHOPPING::getDef('check_github_directory_title');
+      $CLICSHOPPING_Language->loadDefinitions('modules/SecurityCheck/download_directory', null, null, 'Shop');
     }
 
     public function pass()
     {
-      return !is_dir(CLICSHOPPING::getConfig('dir_root', 'Shop') . '.github');
+      $CLICSHOPPING_Template = Registry::get('TemplateAdmin');
+
+      if (DOWNLOAD_ENABLED != 'true') {
+        return true;
+      }
+
+      return is_dir($CLICSHOPPING_Template->getPathDownloadShopDirectory('Private'));
     }
 
     public function getMessage()
     {
-      return CLICSHOPPING::getDef('warning_github_directory_exists', [
-        'github_path' => CLICSHOPPING::getConfig('dir_root', 'Shop') . '.github'
+      return CLICSHOPPING::getDef('warning_download_directory_non_existent', [
+        'download_path' => CLICSHOPPING::getConfig('dir_root', 'Shop') . 'sources/Download/Private/'
       ]);
     }
   }
-