+++ includes/Module/SecurityCheck/extended_last_run.php
@@ -22,7 +22,7 @@
     {
       $CLICSHOPPING_Language = Registry::get('Language');
 
-      $CLICSHOPPING_Language->loadDefinitions('modules/security_check/extended_last_run', null, null, 'Shop');
+      $CLICSHOPPING_Language->loadDefinitions('modules/SecurityCheck/extended_last_run', null, null, 'Shop');
     }
 
     public function pass()