+++ CHANGELOG.md
@@ -5,6 +5,8 @@ update invoice number for pdf
 update bootstrap icon 1.8.3
 update bootstrap@5.2.0-beta1
 small other fix
+Orders : Add a simple order archive batch on 30d,
+
 
 -------------
 Version 3.319