+++ CHANGELOG.md
@@ -6,7 +6,7 @@ update bootstrap icon 1.8.3
 update bootstrap@5.2.0-beta1
 small other fix
 Orders : Add a simple order archive batch on 30d,
-
+fix possible warning DOMDocument::loadHTML(): Tag figure invalid in Entity with ckeditor
 
 -------------
 Version 3.319