+++ CHANGELOG.md
@@ -8,9 +8,7 @@ small other fix
 Orders : Add a simple order archive batch on 30d,
 fix possible warning DOMDocument::loadHTML(): Tag figure invalid in Entity with ckeditor
 Move security check inside Modules
-
-
-
+Move securityPro inside Modules
 
 -------------
 Version 3.319