+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Orders : Add a simple order archive batch on 30d,
 fix possible warning DOMDocument::loadHTML(): Tag figure invalid in Entity with ckeditor
 Move security check inside Modules
 Move securityPro inside Modules
+add ckeditor 5.x
 
 -------------
 Version 3.319